python/currency-converter/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
customtkinter==5.2.2
darkdetect==0.8.0
everapi==0.1.1
freecurrencyapi==0.1.0
idna==3.10
packaging==24.2
pillow==11.1.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.3.0
